Fix for #4701 - OnBeforeTokenRequest can change the URI
